set active 1 for admin, divided controller login into user and registration
